+++ common/include/common/config/GameConfig.h
@@ -113,6 +113,9 @@ struct GameConfig
     CfgVar<bool> af_branding = true;
     CfgVar<bool> world_hud_ctf = false;
     CfgVar<bool> world_hud_overdraw = true;
+    CfgVar<bool> world_hud_big_text = false;
+    CfgVar<bool> world_hud_damage_numbers = true;
+    CfgVar<bool> play_hit_sounds = true;
 
     // Misc
     CfgVar<bool> fast_start = true;